Support callx instruction (interpreter and x86_64 JIT)
Add support for the callx BPF instruction in the interpreter and the
x86_64 JIT.
