+++ vm/ebpf.h
@@ -173,6 +173,7 @@ enum bpf_register
 #define EBPF_OP_JSGE_IMM (EBPF_CLS_JMP | EBPF_SRC_IMM | EBPF_MODE_JSGE)
 #define EBPF_OP_JSGE_REG (EBPF_CLS_JMP | EBPF_SRC_REG | EBPF_MODE_JSGE)
 #define EBPF_OP_CALL (EBPF_CLS_JMP | EBPF_MODE_CALL)
+#define EBPF_OP_CALLX (EBPF_CLS_JMP | EBPF_SRC_REG | EBPF_MODE_CALL)
 #define EBPF_OP_EXIT (EBPF_CLS_JMP | EBPF_MODE_EXIT)
 #define EBPF_OP_JLT_IMM (EBPF_CLS_JMP | EBPF_SRC_IMM | EBPF_MODE_JLT)
 #define EBPF_OP_JLT_REG (EBPF_CLS_JMP | EBPF_SRC_REG | EBPF_MODE_JLT)
